Adds in auto-generated jetty-web xml file for now
Netbeans keeps adding it in automatically. It doesn't actually appear to
be needed to for the web service, and as far as I'm aware it's not used.
I've added a comment to it for now to illustrate this. I could have added
a .gitignore, but it would still be added to the file system by Netbeans,
even if not the repo. I think this is a little clearer.