Adds basic file data logger to the ground station
We use log4j to do this. We set it up to log to a few different files,
depending on the data being logged. We're logging in csv format, as this
should be easy enough to read, and should be easy to import into other
applications.

Also adds console logging to the ground station. This gives us a better
idea of what is going on.

Improves the exception handling in the web service writer. If the PUT
request fails because the web service isn't running, the exception will
be logged rather than crashing the application.