Adds server which can handle the telemetry information which dl-fldigi sends
dl-fldigi sends the data via a PUT request to a RESTful type of url. We
don't know ahead of time what that url will be, so we add a handler for
all requests, regardless of url. dl-fldigi also doesn't specify the
content-type in the http headers, so we force that ourselves to allow
the body-parser module to do its work.

We put the app into online mode, and give it the server address of our
receiver. I've ignored the node dependencies, you should install them
locally.