Updates CameraController to scan selected images dir for new files
When the controller is asked for its current data, it returns the paths
of any files in the images directory. In a future commit, I'll create a
writer which then moves those files into the storage directory for the
flight. Adds unit tests for the camera controller. I've updated the config
properties to work for my current env (Mac). At some point I'll update that
to be correct for the Raspberry Pi, and then I won't commit any more
changes to that path.