Overhaul of how we write data from the balloon
We now have a better idea of how data will be sent between the balloon
and the ground station. We use the telemetryKeys json file as a common
schema for our data, hence why it has been moved up to the top level of
the repository. We now write data as a csv string. Every second, we
query the state of our sensors, and build up a data model based on that.
We then convert that model to a csv string using the data keys in the
schema.

Camera data is different from other data, so I no longer try and treat it
in the same way. As well as our other sensors, we now have a separate
camera sensor and data writer, with their own APIs.