Creates a separate database server
Moves any database operations out of the dl-fldigi receiver and into this
new server. When the dl-fldigi receiver receives new data, it POSTs it to
the database server. This decouples dl-fldigi from our database
implementation. If we were to move to a different implementation, we
shouldn't have to change the dl-fldigi receiver.