Initial draft of the ground station web app
Takes the Angular seed project, and updates it with html taken from
Bootstrap. I've then renamed the first page to /telemetry, and renamed
the associated controller accordingly. Updates the page so that it displays
something, even if that's just to show that it doesn't yet have any data.
In the next commit, I'll update it so it picks up data from the database
server.