Updates loggers to initialise from config file
This more closely follows the guidance given in the log4j documentation.
We add a logger.properties file to the project, which we then load in at
the start. The log4j PropertyConfigurator class configures the root logger
based on this config. We default the root logger to append to the console
at the moment. This gets rid of warnings we were getting where some loggers
in the hierarchy didn't have an appender set.