Adds simple web page to display latest webcam images
It should be easy for us to inspect the latest webcam images during
testing, so I've created this little web page to allow that. It refreshes
the page at regular intervals. This change updates the webcam image
capture script so that it copies the image it captures into the display
folder for inspection.

I've ignored the webcam.jpg file in Git, as this will be continuously
updated.