Updates GPS sensor controller so it catches native linker errors
Pi4j wraps some native libraries which help to interface with the
hardware on the Pi. If we run the pi4j libraries on machines other than
the Pi, we get native linker errors where these native libraries aren't
resolved. In the case where we aren't on the Pi, we can't really
communicate with the GPS module anyway, so for now we're catching these
exceptions and returning an empty data object. This allows the app to
keep running without falling over in a heap.