Adds first version of serial port data writer
Uses the Pi4J library. When we create the writer, we also create a
reader to echo out any messages written. This is used as a check that data
is written correctly to the serial port.

In the balloon controller code, we try and create a serial port writer. If
that fails (errors thrown from the constructor), we fall back to using
HTTP. This code is looking a bit messy now, so in the next commit I'll look
and pulling out the writer and sensor creation in a better way.