Adds scripts to allow us to run jar as root
To run as root, we can no longer run directly from Netbeans - we'll need
to go from the command line instead. To help with this, I've created
scripts to build our jar, copy it to the Pi, and then run the jar as root
on the Pi.

These scripts depend on Ant being installed on the target machine, and your
public RSA key being added to the authorized_keys on the Pi.