Adds transistor switch controller into the BalloonController
It is the first implementation of a DataModelConsumer. It reads the height
from the model and processes it.

Also deletes the project.latex project, and moves its classes into the
project.latex.balloon class. Having a few classes in a separate project
didn't seem to make any sense any more.